+++ compiler-core/src/error.rs
@@ -220,6 +220,11 @@ fn did_you_mean(name: &str, options: &[String], alt: &'static str) -> String {
 }
 
 impl Error {
+    pub fn pretty_string(&self) -> String {
+        let mut nocolor = Buffer::no_color();
+        self.pretty(&mut nocolor);
+        String::from_utf8(nocolor.into_inner()).expect("Error printing produced invalid utf8")
+    }
     pub fn pretty(&self, buf: &mut Buffer) {
         use crate::type_::Error as TypeError;
         use std::io::Write;
@@ -731,14 +736,14 @@ also be labelled.",
                     .unwrap();
                 }
 
-                TypeError::UnknownField {
+                TypeError::UnknownRecordField {
                     location,
                     typ,
                     label,
                     fields,
                 } => {
                     let diagnostic = Diagnostic {
-                        title: "Unknown field".to_string(),
+                        title: "Unknown record field".to_string(),
                         label: did_you_mean(label, fields, "This field does not exist"),
                         file: path.to_str().unwrap().to_string(),
                         src: src.to_string(),
@@ -749,7 +754,7 @@ also be labelled.",
 
                     writeln!(
                         buf,
-                        "The value has this type:
+                        "The record being updated has this type:
 
 {}
 ",
@@ -761,7 +766,7 @@ also be labelled.",
                         writeln!(buf, "It does not have any fields.",).unwrap();
                     } else {
                         write!(buf, "It has these fields:\n\n").unwrap();
-                        for field in fields {
+                        for field in fields.iter().sorted() {
                             writeln!(buf, "    .{}", field).unwrap();
                         }
                     }
@@ -1401,7 +1406,7 @@ and try again.
 
                     writeln!(
                         buf,
-                        "You are attempting to update a record using a value that is not a record constructor",
+                        "Only record constructors can be used with the update syntax.",
                     )
                     .unwrap();
                 }